requirements.txt
# AION Dependencies
# =================

# Core
openai>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0

# Rich CLI Styling (optional aber recommended)
rich>=13.0.0

# Web UI
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# TTS (optional — engine=edge empfohlen, engine=sapi5 als Fallback)
edge-tts>=6.1.0

# File processing (Web UI uploads)
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Browser automation (playwright_browser plugin)
# After install: playwright install chromium
playwright>=1.40.0

# Messaging channels (optional — je nach gewünschten Kanälen)
discord.py>=2.3.0          # Discord-Bot (discord_bot plugin)
slack-bolt>=1.18.0         # Slack-Bot (slack_bot plugin)
